+++ packages/virtual/core/src/index.ts
@@ -1,3 +1,8 @@
+/**
+ * @packageDocumentation
+ * {@inheritDoc VirtualScroller}
+ */
+
 export { default as VirtualScroller } from "models/VirtualScroller";
-export { Event, ALL_EVENTS } from "constants/";
+export { VirtualScrollerEvent, _ALL_EVENTS } from "constants/";
 export * from "./types";