+++ packages/virtual/react/src/components/List/index.tsx
@@ -1,10 +1,16 @@
 import { mapVisibleRange } from "utils/rangeMappers";
 import useSyncedStyles from "hooks/useSyncedStyles";
 import Subscription from "components/Subscription";
-import { Event } from "@af-utils/virtual-core";
+import { VirtualScrollerEvent } from "@af-utils/virtual-core";
 import type { ElementType } from "react";
 import type { ListProps } from "types";
 
+/**
+ * @public
+ * React component.
+ * {@link ListProps}.
+ * Small abstraction, which in 90% cases allows to avoid hook boilerplate.
+ */
 const List = <T extends ElementType = "div">({
     model,
     children: Item,
@@ -34,7 +40,10 @@ const List = <T extends ElementType = "div">({
             {header}
             <div ref={outerRef}>
                 <div ref={innerRef}>
-                    <Subscription model={model} events={[Event.RANGE]}>
+                    <Subscription
+                        model={model}
+                        events={[VirtualScrollerEvent.RANGE]}
+                    >
                         {() =>
                             mapVisibleRange(model, i => (
                                 <Item