+++ packages/virtual/react/src/components/Subscription/index.tsx
@@ -1,6 +1,12 @@
 import useComponentSubscription from "hooks/useComponentSubscription";
 import type { SubscriptionProps } from "types";
 
+/**
+ * @public
+ * React component.
+ * Rerenders only on certain {@link @af-utils/virtual-core#(VirtualScrollerEvent:variable)}.
+ * Allows to optimize performance.
+ */
 const Subscription = (props: SubscriptionProps) => {
     useComponentSubscription(props.model, props.events);
     return props.children();