+++ packages/virtual/react/src/hooks/useScroller/index.ts
@@ -1,10 +1,12 @@
 import useIsomorphicLayoutEffect from "hooks/useIsomorphicLayoutEffect";
-import type { VirtualScroller } from "@af-utils/virtual-core";
+import type { ScrollElement, VirtualScroller } from "@af-utils/virtual-core";
 
-const useScroller = (
-    model: VirtualScroller,
-    scroller: Parameters<VirtualScroller["setScroller"]>[0]
-) =>
+/**
+ * @public
+ * React hook.
+ * Synchronizes scroller with model. Should be used in window-scroll cases, otherwise `ref={model.setScroller}` is preferrable.
+ */
+const useScroller = (model: VirtualScroller, scroller: ScrollElement | null) =>
     useIsomorphicLayoutEffect(() => {
         model.setScroller(scroller);
         return () => model.setScroller(null);