+++ packages/virtual/react/src/hooks/useVirtualModel/index.ts
@@ -1,8 +1,19 @@
 import { useRef } from "react";
-import { VirtualScroller } from "@af-utils/virtual-core";
+import {
+    VirtualScroller,
+    VirtualScrollerInitialParams
+} from "@af-utils/virtual-core";
 
-const useVirtualModel = (
-    params: ConstructorParameters<typeof VirtualScroller>[0]
-) => (useRef<VirtualScroller>().current ||= new VirtualScroller(params));
+/**
+ * @public
+ * React hook.
+ * Creates and stores one {@link @af-utils/virtual-core#VirtualScroller} instance.
+ * It does not get recreated during component lifecycle.
+ * Normally {@link useVirtual} should be used.
+ * @param params - {@link @af-utils/virtual-core#VirtualScrollerInitialParams}
+ * @returns same {@link @af-utils/virtual-core#VirtualScroller} instance
+ */
+const useVirtualModel = (params: VirtualScrollerInitialParams) =>
+    (useRef<VirtualScroller>().current ||= new VirtualScroller(params));
 
 export default useVirtualModel;