+++ packages/website/components/Menu/index.tsx
@@ -7,8 +7,7 @@ import startCase from "utils/startCase";
 const DOCS_STRUCTURE = [
     ["Getting started", ""],
     ["FAQ", "/faq"],
-    ["Headless", "/headless"],
-    ["List", "/list"],
+    ["Reference", "/reference"],
     ["Bundle size impact", "/size"]
 ] as const;